\name{SCBmeanfd-package}
\alias{SCBmeanfd-package}
\alias{SCBmeanfd}
\docType{package}
\title{
Simultaneous Confidence Bands for Mean Functions}
\description{
Statistical methods for estimating and inferring the mean of functional data. The methods include simultaneous confidence bands, local polynomial fitting,  bandwidth selection by plug-in and cross-validation, goodness-of-fit tests for parametric models, equality tests for two-sample problems, and plotting functions.}
\details{
\tabular{ll}{
Package: \tab SCBmeanfd\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2016-09-24\cr
License: \tab GPL-3\cr
}

}
\author{
David Degras <ddegrasv@gmail.com>
}

\references{
Azzalini, A. and Bowman, A. (1993). On the use of nonparametric regression for checking linear relationships. \emph{Journal of the Royal Statistical Society. Series B} \bold{55}, 549-557.

Benhenni, K. and Degras, D. (2014). Local polynomial estimation of the mean function and its derivatives based on functional data and regular designs. \emph{ESAIM: Probability and Statistics} \bold{18}, 881-899.

Degras, D. (2011). Simultaneous confidence bands for nonparametric regression with functional data. \emph{Statistica Sinica} \bold{21}, 1735-1765.

Rice, J. A. and Silverman, B. W. (1991). Estimating the mean and covariance structure nonparametrically when the data are curves. \emph{Journal of the Royal Statistical Society. Series B} \bold{53}, 233-243.
}

\keyword{ package }


