\name{SCBmeanfd-package}
\alias{SCBmeanfd-package}
\alias{SCBmeanfd}
\docType{package}
\title{
Simultaneous Confidence Bands for Mean Functions}
\description{
This package implements statistical methods for the mean of functional data based on local polynomials and simultaneous confidence bands (SCB). Functions for kernel bandwidth selection, SCB estimation, goodness-of-fit tests, and two-sample tests are provided. Visualization tools are also available. 
}
\details{
\tabular{ll}{
Package: \tab SCBmeanfd\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-01-29\cr
License: \tab GPL-3\cr
}

}
\author{
David Degras <ddegrasv@depaul.edu>
}

\references{
Azzalini, A. and Bowman, A. (1993). On the use of nonparametric regression for checking linear relationships. \emph{Journal of the Royal Statistical Society. Series B} \bold{55}, 549-557.

Benhenni, K. and Degras, D. (2011). Local polynomial estimation of the average growth curve with functional data. \url{http://arxiv.org/abs/1107.4058}

Degras, D. (2011). Simultaneous confidence bands for nonparametric regression with functional data. \emph{Statistica Sinica} \bold{21}, 1735-1765.

Rice, J. A. and Silverman, B. W. (1991). Estimating the mean and covariance structure nonparametrically when the data are curves. \emph{Journal of the Royal Statistical Society. Series B} \bold{53}, 233-243.
}

\keyword{ package }


